# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Atoma API
  description: |
    REST API for Atoma - a minimalist iOS app for habit tracking, task management, and budget control.

    **Tagline**: habits, tasks, money - in one flow

    ## Authentication
    All endpoints (except auth) require a valid JWT access token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Response Format
    All successful responses follow this format:
    ```json
    { "data": { ... } }
    ```

    Error responses:
    ```json
    { "error": { "code": "ERROR_CODE", "message": "..." } }
    ```
  version: 1.0.0
  contact:
    name: Atoma Support
    url: https://api.azamatbigali.online
  license:
    name: Private
servers:
  - url: https://api.azamatbigali.online/api/v1
    description: Production server
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /habits:
    get:
      tags: [Habits]
      summary: List Habits
      description: Get all habits for the authenticated user
      operationId: listHabits
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of habits
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Habit'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Habits]
      summary: Create Habit
      description: Create a new habit
      operationId: createHabit
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHabitRequest'
      responses:
        '201':
          description: Habit created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Habit'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /habits/{id}:
    get:
      tags: [Habits]
      summary: Get Habit
      description: Get a single habit by ID
      operationId: getHabit
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Habit details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Habit'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Habits]
      summary: Update Habit
      description: Update an existing habit
      operationId: updateHabit
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateHabitRequest'
      responses:
        '200':
          description: Habit updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Habit'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Habits]
      summary: Delete Habit
      description: Delete a habit
      operationId: deleteHabit
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Habit deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  parameters:
    idParam:
      name: id
      in: path
      required: true
      description: Resource UUID
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: Invalid or expired token
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: VALIDATION_ERROR
              message: Title is required
...
  schemas:
...
    CreateHabitRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 200
        goal_id:
          type: string
          format: uuid
        icon:
          type: string
          default: "🎯"
        color:
          type: string
          default: green
        period:
          type: string
          enum: [daily, weekly, monthly]
          default: daily
        target_value:
          type: integer
          description: Target for measurable habits
        unit:
          type: string
          description: Unit for target (e.g., pages, minutes)
    UpdateHabitRequest:
      type: object
      properties:
        title:
          type: string
        goal_id:
          type: string
          format: uuid
        icon:
          type: string
        color:
          type: string
        period:
          type: string
          enum: [daily, weekly, monthly]
        target_value:
          type: integer
        unit:
          type: string
        archived_at:
          type: string
          format: date-time
          description: Set to archive, empty string to unarchive
...
    Habit:
      type: object
      properties:
        id:
          type: string
          format: uuid
        goal_id:
          type: string
          format: uuid
        title:
          type: string
        icon:
          type: string
        color:
          type: string
        period:
          type: string
          enum: [daily, weekly, monthly]
        completed_dates:
          type: array
          items:
            type: string
            format: date
        target_value:
          type: integer
        unit:
          type: string
        progress_values:
          type: object
          additionalProperties:
            type: integer
        streak:
          type: integer
          description: Current streak (calculated by backend)
        created_at:
          type: string
          format: date-time
        archived_at:
          type: string
          format: date-time

    # ==================== TASKS ====================
...
    MessageResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            message:
              type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...